requirements.txt
python-chess
zstandard
